Added # of distinct ETH addresses to tooltip.
This new addition handles for the rare case where multiple ETH addresses submit orders of equal price and equal volume. For example, if 2 users both submit orders at a price of $1000 for 250 ETH, this would presently show up as a bubble of 500 ETH for $1000, but that is partially misleading. Now, the user can hover over and, thanks to the tool-tip, see the exact number of distinct addresses.